WIP: begin v2 Penumbra support in Hermes
This starts constructing a v2 of Penumbra support in Hermes, using the
multi-chain backend support we added.

We should be careful to write this code so that it's _maintainable_ as a set of
changes on top of upstream Hermes.  Thus, whenever I've encountered things that
should change about the Hermes code generally, I've marked them with
`TODO(extract)`. That way, we can note changes we'd like to make / need to make
along the way, send them upstream as small, self-contained PRs, and then rebase
this work once they're merged.

The architecture we upstreamed should allow us to keep all the Penumbra-related
code in a submodule, isolated from the rest of the Hermes code, so that future
rebasing won't generate conflicts -- and ideally we could even merge support
upstream, though that will probably need to wait until we can publish all the
Penumbra crates.