Update penumbra deps to v80 (#48)
Bumping the Penumbra dependencies to v0.80.x, specifically on the current main branch of the protocol repo. Ideally we'd tag a v0.80.10, containing fixes from penumbra-zone/penumbra#4905, but we haven't done that just yet.

Notably, the bump in Penumbra protocol deps will involve a schema change for the view database. I haven't tested locally yet, but I expect Hermes when upgraded to throw an error about the view schema mismatch. If and when that happens during interactive testing on the testnet, I'll make sure to document those errors so we can communicate the need to perform maintenance to operators.